simplify autosize hack directive
This commit will add a simplified version of the autosize hack directive.
This directive now works with the 'autosize' attribute/input of  the
MatDrawerContainer and MatSidenavContainer selectors.
Its role is to simply remove the animation class added to the
container when a drawer is toggled open or closed.
The functionality to animate the sidenav content has been moved
to the ExpansionViewport component.